requirements.txt
appdirs==1.4.4
asn1crypto==0.24.0
biopython==1.77
bleach==3.1.2
cryptography==2.6.1
decorator==4.3.0
defusedxml==0.5.0
distlib==0.3.1
distro-info==0.21
entrypoints==0.3
filelock==3.0.12
html5lib==1.0.1
importlib-metadata==1.7.0
ipykernel==4.9.0
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
keyring==17.1.1
keyrings.alt==3.1.1
MarkupSafe==1.1.0
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.8
pandocfilters==1.4.2
pbr==5.4.5
pexpect==4.6.0
pickleshare==0.7.5
prometheus-client==0.6.0
prompt-toolkit==1.0.15
pycrypto==2.6.1
pycurl==7.43.0.2
Pygments==2.3.1
PyGObject==3.30.4
python-apt==1.8.4.1
python-dateutil==2.7.3
pyxdg==0.25
PyYAML==5.3.1
pyzmq==17.1.2
requests==2.24.0
SecretStorage==2.3.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.12.0
stevedore==2.0.1
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
unattended-upgrades==0.1
virtualenv==20.0.26
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==2.0.0
zipp==3.1.0
